requirements.txt
langchain
langchain-core  
uvicorn
python-dotenv
langchain-google-genai 
fastapi 
langserve 
sse_starlette